Unit tests for a more complex database
Added a more complex database and some unit tests for it. More tests for this database are needed since confirming functionality here confirms the correctness of the engine more than the unit tests for the simpler data base. The unit tests there should be more regarded as ensuring the reliability of the engine.